Better handling of $refs (closes #13)
$ref handling during get_schema() is now done by the Ref.descend_scema().

References are followed during schema descend, but are retained in the final
output. This behavior is now consistent. Previously, most $refs in the output
were replaced by the actual object they were referencing, except for the
arrayMergeById strategy, which retained $refs in its output.

This also obsoletes the need for WalkSchema.resolve_refs() calls in strategies.