Allow any strategy to be used with anyOf, etc.
Make it possible to explicitly define any strategy at the same level as anyOf,
etc. JSON Schema allows other keywords to be defined at the same level as
anyOf, allOf, oneOf, etc. So it makes sense that a merge strategy can be
defined at this level as well.

We still raise an error if a strategy is not explicitly defined for anyOf,
allOf for backwards compatibility and protection against mistakes, since
implicit objectMerge might be confusing in this case.