Remove "meta" argument throughout the code.
The "meta" argument was just a special case cludge to pass an option to the
version strategy and had to be passed explicitly throughout most of the code.
mergeOptions is a better and more general way for doing this,

With this change, version strategy gets two new mergeOptions: metadata and
metadataSchema. They behave exactly like the meta arguments did to the
Merger.merge() and Merger.get_schema() methods.

To allow for the use case where a different metadata needs to be passed to each
merge()/get_schema() call without changing the schema, the new merge_options
argument adds global options to strategies.